Унификация заголовков секций и исправление дизайн-системы
- Добавлена фабрика BreezSectionHeader.settings() для настроек
- Удалён BreezSectionTitle (заменён на единый BreezSectionHeader)
- Исправлены hardcoded значения в диалогах (Colors.transparent, размеры, opacity)
- Переработан DeleteConfirmDialog на стандартные компоненты
- Исправлен schedule_screen: AlertDialog → Dialog + BreezSectionHeader.dialog
- Добавлено поле macAddress в UnitState
- Стандартизированы BreezDialogButton (icon, height, danger style)
- Исправлены boolean параметры в operation_graph (named parameters)

+++ lib/presentation/screens/dashboard/dialogs/add_unit_dialog.dart
@@ -10,7 +10,6 @@ import 'package:hvac_control/core/theme/app_theme.dart';
 import 'package:hvac_control/core/theme/spacing.dart';
 import 'package:hvac_control/generated/l10n/app_localizations.dart';
 import 'package:hvac_control/presentation/widgets/breez/breez.dart';
-import 'package:hvac_control/presentation/widgets/breez/breez_list_card.dart';
 
 // =============================================================================
 // CONSTANTS
@@ -96,7 +95,7 @@ class _AddUnitDialogState extends State<AddUnitDialog> {
     final l10n = AppLocalizations.of(context)!;
 
     // Адаптивная ширина: максимум как форма регистрации
-    final maxWidth = min(MediaQuery.of(context).size.width - 48, AuthConstants.formMaxWidth).toDouble();
+    final maxWidth = min(MediaQuery.of(context).size.width - 48, AuthConstants.formMaxWidth);
 
     return Dialog(
       backgroundColor: colors.card,
